third_party/blink/web_tests/platform/generic/svg/css/svg-attribute-length-parsing-expected.txt

Test CSS parsing on SVG presentation attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(text, null).fontSize is "16px"
PASS document.defaultView.getComputedStyle(text, null).fontSize is "100px"
PASS document.defaultView.getComputedStyle(text, null).fontSize is "16px"
PASS document.defaultView.getComputedStyle(text, null).fontSize is "16px"
PASS successfullyParsed is true

TEST COMPLETE
